go: Use a single go.mod
This solves a few issues with the multi-module repo.
Using one go.mod means that changing gobpfd immediately gets picked up
in bpfd-operator, or the examples. It also makes CI simpler and things
like go build/go lint etc..

The downside is that the module graph for gobpfd is huge.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>