move the rpc server to it's own binary
This is a huge commit:

- Move the GRPC serve out into it's own binary that
simply calls into the core bpfman library. This
allows us to fully stop acting like a long
running daemon even for the kubernetes and non-rust
language use cases. This binary lives in the
existing bpfman-api crate, making this work required
removing a bunch of circular dependencies where
there were internal types defined in bpfman-api
that were being used in bpfman.

- remove some unused deps.

- fix log levels in k8s integration tests Make sure
we are logging at the debug level in both
the daemon and operator when running our kubernetes
integration tests.

- Now that bpfman is a library and the rpc server
has been extracted we need to make sure
that the sled database lock is released whenever
it's not being used so other processes can call the
library on the same host. This wan't possible with
the root database reference as a lazy static variable
since once it was initialized by a process it would
never be dropped until the calling process exited.
This is an issue when the calling process is our
long lived rpc shim. To fix this simply add the
root database instance as a member of the BpfManager
and then pass it via arguments where ever it's needed.
Additionally fixup the retry mechanism for acquiring
the lock on the filesystem backed sled db instance
to default accordingly and tweak it to retry a bit
more when running in kubernetes.

Signed-off-by: Andrew Stoycos <astoycos@redhat.com>