multiarch: Build multiarch binaries and images
Since bpfman-operator is moving to it's own repo, no longer build and
push the bpfman-operator and bpfman-agent from the bpfman repo, but
instead, build and push from the bpfman-operator repo.

Rework the Makefile such that it no longer loops through each
architecture. Building all architectures is done in github actions. The
Makefile still allows local cross compiling for a single architecture,
just not all architectures in one command. This simplifies the Makefile.

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>