bpfman-operator: Add test helper for BpfmanConfigReconciler setup
Introduce a new test helper function `setupTestEnvironment` to
standardise the setup process for BpfmanConfigReconciler tests. This
function handles the initialisation of the fake client, the registration
of necessary types with the runtime scheme, and the configuration of the
reconciler.

The helper enables specific OpenShift testing by conditionally setting
the RestrictedSCC field on the reconciler object.

The logic is taken from the existing TestBpfmanConfigReconcileAndDelete
function. The previous setup logic duplicated paths to various YAML
files by pre-pending ../../. This new setup function dispenses with the
duplicated paths by using the constants declared in the internal package
and resolves the path at runtime to the actual files in the filesystem
relative to the test directory. This allows us to refer to the canonical
path without having to duplicate it, ensuring that if we ever change the
filename then the test code doesn't need to be updated.

This commit allows us to share the test setup logic so that we can
subsequently modify TestBpfmanConfigReconcileAndDelete to verify that
the Restricted SCC is created in the OpenShift test case, which was
previously not tested.

Signed-off-by: Andrew McDermott <amcdermo@redhat.com>