Handle deletes of BpfPrograms for BpfApplications
When BpfPrograms are deleted by unLoadAndDeleteBpfProgramsList(), the
relevant controller needs to know that the programs are being deleted so
it can update the global status correctly. This change updates the
status to bpfmaniov1alpha1.BpfProgCondUnloaded before deleting the
program object, which makes it work more like the delete case of
reconcileBpfProgram().

Signed-off-by: Andre Fredette <afredette@redhat.com>