Add separate workflow to run Kubernetes integration tests
Runs for pull requests, on demand, and periodically.

This allows us to easily run an integration test on a new version of
bpfman, and also runs the integration tests periodically so that they
run with the latest bfpman even if there are no changes in there are no
changes in bpfman-operator.

Signed-off-by: Andre Fredette <afredette@redhat.com>