WIP: Initial pass at BpfApplication CRD update for load/attach split
This still includes the attach info with the programs, but it is
optional, and may be updated later.  This allows the program to be
loaded before any attachments are made, and allows attachments to be
added after the program has been loaded.

TODO: Add a per-node CRD (BpfProgram analog) to maintain the per-node
state for the BpfApplication.

Signed-off-by: Andre Fredette <afredette@redhat.com>