docs: move bpfman-operator details from README to docs
The bpfman-operator README.md contained operator details that belonged
in the general docs. Move this content to the docs, and expand on the
details where needed. Also update the bpfman-operator README to be
closer to the bpfman README by adding the bpfman logo, CNCF logo,
some of the relavent badges, and the star history.

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>