remove dependency on security-profiles-operator
After deploying the selinux profile, the status on the Selinux Profile
is “Pending”. security-profiles-operator is currently deployed in
OpenShift by making it a dependency of bpfman-operator. As a result, the
security-profiles-operator is deployed in the bpfman namespace.
security-profiles-operator encounters issues with this because there are
other daemonsets in the namespace. Short term, remove the dependency.
security-profiles-operator is still required, it just won't be
auto-installed.

Related: #331
Related: https://github.com/kubernetes-sigs/security-profiles-operator/issues/2699

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>