update app integration test result processing
The TestApplicationGoCounter in app_test.go is a little flaky,
occasionally failing. Seems to be failing more often with the recent
changes in bpfman. Turns out the log processing was always looking at
the oldest retrieved log, so if it didn't pass on the first read, it
would never pass. It was just random timing if the first log contained
packet counts or not.

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>