Change ContainerSelector to NetworkNamespaceSelector for XDP, TC, and TCX
The NetworkNamespaceSelector includes a pod selector, and for the
cluster-scoped CRD, it also includes an optional namespace.  When used,
bpfman will attach the program inside the network namesapces of the pods
selected.

/run/netns from the host is mounted at /host/netns in the bpfman
container. Which supports interface discovery which is implemented in
https://github.com/bpfman/bpfman-operator/pull/408 and alows the bpfman
agent to attach interfaces disovered in network namespaces found in
/run/netns.

Fixes: #410

Signed-off-by: Andre Fredette <afredette@redhat.com>