api review: update field descriptions
As part of the API review, a comment was made to improve the description
of all fields. This commit makes a pass at the ClusterBpfApplication CRD
fields. Once acceptable, the chnages will be propgated to
BpfApplication, ClusterBpfApplicationState and BpfApplicationState. Many
of the fields are the same or similiar, so won't general consensus
first.

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>