Check for Duplicate Network Namespace Inodes
It’s possible for multiple pods to share the same network namespace,
even though unique handles for that namespace are created
in /var/run/netns. This most commonly occurs when pods are configured
with "hostNetwork": true.

The interface discovery process watches /var/run/netns, and when it
detects a new entry, it begins monitoring the interfaces in that
namespace and informs bpfman. However, if bpfman attempts to attach
multiple programs — especially TCX programs — to the same interface
in the same netns, it will return an error. It's also possible to run
into this situation using the network namespace selector.

For TC and XDP programs, the attach might not fail immediately, but
bpfman limits the number of programs per attach point (currently to
10). If more than 10 handles reference the same network namespace —
which is common on OpenShift clusters — this limit is quickly
reached. That can lead to subtle and unexpected failures, even though
the individual netns paths appear different.

The solution is to compare inodes rather than file paths when
identifying network namespaces. This allows the system to correctly
recognize when multiple entries actually refer to the same underlying
namespace.

Signed-off-by: Andre Fredette <afredette@redhat.com>