Fix issues attaching inside network namespaces
bpfman-agent needs to mount /proc so it can get the inodes of network
namspaces with paths in /proc.

I've also run across some cases where network namespaces that don't
exist appear in the list of discovered interfaces.  This might be
possible on a transient basis, but it didn't seem like they were
getting cleaned up, and this results in a permanent error for bpfman
because it tries to attach a program inside a network namespace that
doesn't exist which results in a failure.  If this happens, the code
now logs an error and doesn't try to attach the link.

Signed-off-by: Andre Fredette <afredette@redhat.com>