Improve bundle versioning using VERSION file
Move version information into a standalone VERSION file to provide a
clear, central source of truth for the project version and future
runtime consumers (e.g. CLI programs that may support a --version
flag).

There's nothing inherently wrong with the previous approach of
hard-coding VERSION in the Makefile. The real issue is that we had no
process for bumping the version immediately after a release tag was
cut. As a result, subsequent development continued to build and push
bundle and catalog images using the same version tag.

For example, once we had tagged 0.5.6, any further updates to the
bundle--such as updated manifests--were still pushed to
`bpfman-operator-bundle:v0.5.6`. This isn't expected behaviour and
violates the assumption that tagged images are stable and immutable.

Signed-off-by: Andrew McDermott <amcdermo@redhat.com>