docs: add script to run kubectl explain on all crds
Add a script that given the root CRD name, will run `kubectl explain` on
all <Object> and sub-<Object> fields and pipe the output to a txt file
that is checked in. Also add `make explain` to the Makefile to automate
the running of the script. This serves two purposes:
1) Documents the CRDs in an easily readable format for users of the CRD.
2) Currently needs to be manually run, but shows how changes to the API
   effect the CRDs. This could be automated in the github workflow if
   the CRDs are loaded into a Kubernetes cluster in github actions.

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>